Add --rev-list optional argument
By default, git-burn uses git rev-list "$old..$new" --not --branches
--tags to obtain the list of commits to lint.

This is fine for running inside update hook, as the list returned
is the set of _new_ commits just pushed, and excludes commits
already reachable from existing tags and branches [1].

However, this is not flexible enough if tool is to be run manually.
For example, if one wants to lint local branch, the rev-list
is something like "master..HEAD". Note that in this scenario --not
--branches would exclude everything since everything is reachable from
the currently checked out branch.

[1] https://stackoverflow.com/questions/35383560/best-way-to-process-only-new-commits-in-an-update-hook